fold LikelihoodUtil into LikelihoodSuite
